Add more description, FW and autohit feature.
Slightly alter the appearence of the website. Add an info text with
valid formats for each of the stats.
Add a Flesh Wound feature for lowering the roll requirements on injury
rolls. Fix the functionality of being able to enter 'D6' or '2D6'
attacks or damage. Implement autohit with putting an 'a' into the field
WS/BS.